requirements.txt
tensorflow-gpu==2.8.0
opencv-python==4.5.5.62
Flask==2.0.2